Remove variable assignment so output is valid JSON
We want to be able to read the JSON back in to convert it to TXF for
TurboTax desktop to consume, so let's make the converter output a valid
JSON object. This requires the user to type a variable assignment
(`entries = ...`) into the JS console when using TurboTax Web, but that
doesn't seem like a big deal.